+++ src/event_loop/background_thread.rs
@@ -6,7 +6,7 @@
 use anymap::Entry;
 use crossbeam::channel;
 use parking_lot::Mutex;
-use std::sync::{Arc, Weak};
+use std::sync::{Arc, LazyLock, Weak};
 use std::thread::{self, JoinHandle};
 
 use super::MainThreadExecutor;
@@ -73,10 +73,8 @@ where
 // Rust does not allow us to use the `T` and `E` type variable in statics, so this is a
 // workaround to have a singleton that also works if for whatever reason there arem ultiple `T`
 // and `E`s in a single process (won't happen with normal plugin usage, but sho knwos).
-lazy_static::lazy_static! {
-    static ref HANDLE_MAP: Mutex<anymap::Map<dyn std::any::Any + Send>> =
-        Mutex::new(anymap::Map::new());
-}
+static HANDLE_MAP: LazyLock<Mutex<anymap::Map<dyn std::any::Any + Send>>> =
+    LazyLock::new(|| Mutex::new(anymap::Map::new()));
 
 impl<T: Send + 'static, E: MainThreadExecutor<T> + 'static> WorkerThread<T, E> {
     fn spawn() -> Self {